requirements.txt
fastapi
uvicorn
python-socketio
aiofiles
shortuuid